fetch questions around and see them in questions tab

+++ EtrafaSor/EtrafaSorHTTPRequestHandler.h
@@ -28,7 +28,7 @@
                                         sender:(id<EtrafaSorHTTPRequestHandlerDelegate>)sender;
 
 + (void)questionsAroundCenterCoordinate:(CLLocationCoordinate2D)coordinate
-                             withRadius:(CGFloat)radius
+                             withRadius:(NSInteger)radius
                                    user:(Profile *)user
                                  sender:(id<EtrafaSorHTTPRequestHandlerDelegate>)sender;
